requirements.txt
Django==2.2.7
django-environ==0.4.5
djangorestframework==3.11.0
mysql-connector-python==8.0.18
protobuf==3.11.2
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
